set the fallback log creator up instead of the log creator
The reason for this change is to be able to use another logging mechanism (slf4j) when available on the classpath.  The default is the Ant logger.
`LogCreatorFactory.getLogCreator` uses the class loader from *context* classloader which is initialized with `flyway.classpath` as a classpath.
When Flyway finds a well-known logging library initializes a new log creator (for example `Slf4jLogCreator` for `slf4j`).
The creator's method `logCreator.createLogger` doesn't use the context classloader anymore, it means it initializes the logger from the task classloader itself (defined in `<taskdef>`) and it means **the logging libraries must be on the _task_ classpath**.

*Example*: To enable slf4j with log4j2:
- Copy flyway and logging libraries into `flyway.lib.path` used in the `taskdef`.
- Set a configuration via environment or system properties (eg. `ANT_OPTS=-Dlog4j.configurationFile=/path/to/log4j2.xml`) if it is not included in a jar on the classpath yet.